requirements.txt
openai-whisper
setuptools-rust
# ffmpeg must be installed on the system (e.g. sudo apt install ffmpeg)
